split transaction routes on frontend
